+++ aldrin/CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Discoverer::entries()`, which returns a `DiscovererEntries` to iterate over all entries.
 - Add a method `service()` returning the service id to `Call`, `Promise`, `UnknownCall`,
   `low_level::Call` and `low_level::Promise`.
+- Add a method `service()` returning the service id to `Event`, `UnknownEvent` and
+  `low_level::Event`.
 
 ### Changed